sm_nav2_gazebo_test_3: aim the collision-abort finale at the south wall
The pillar target proved too fragile: it subtends ~13 degrees at 1.5 m, and
the accumulated odometry drift from the dead-reckoned primitive laps (spin
wheel slip especially) let the approach drive past it - the run validated
the success path instead of the abort path. The south wall spans the whole
arena, so any southward heading intersects it regardless of drift. Also caps
the behavior server rotational velocity at 0.5 rad/s: fast pure rotation
visibly slips the TB3 wheels in gazebo.
