+++ smacc2_sm_reference_library/sm_nav2_gazebo_test_3/config/nav2_params.yaml
@@ -408,9 +408,11 @@ behavior_server:
     robot_base_frame: base_link
     transform_tolerance: 0.1
     simulate_ahead_time: 2.0
-    max_rotational_vel: 1.0
-    min_rotational_vel: 0.4
-    rotational_acc_lim: 3.2
+    # slower spin cap: fast pure rotation slips the TB3 wheels in gazebo and the
+    # accumulated odometry drift is clearly visible over repeated 360s
+    max_rotational_vel: 0.5
+    min_rotational_vel: 0.2
+    rotational_acc_lim: 1.5
 
 waypoint_follower:
   ros__parameters:
