+++ smacc2_sm_reference_library/sm_nav2_gazebo_test_3/include/sm_nav2_gazebo_test_3/states/st_back_up_safe.hpp
@@ -25,7 +25,7 @@ using namespace smacc2::default_transition_tags;
 
 // STATE DECLARATION
 //
-// Collision-abort demo, step 3: back away from the pillar to a safe distance
+// Collision-abort demo, step 3: back away from the wall to a safe distance
 struct StBackUpSafe : smacc2::SmaccState<StBackUpSafe, SmNav2GazeboTest3>
 {
   using SmaccState::SmaccState;
@@ -49,7 +49,7 @@ struct StBackUpSafe : smacc2::SmaccState<StBackUpSafe, SmNav2GazeboTest3>
 
   void onEntry()
   {
-    RCLCPP_INFO(getLogger(), "StBackUpSafe: onEntry() - backing away from the pillar");
+    RCLCPP_INFO(getLogger(), "StBackUpSafe: onEntry() - backing away from the wall");
   }
 };
