+++ smacc2_sm_reference_library/sm_nav2_gazebo_test_3/include/sm_nav2_gazebo_test_3/states/st_drive_at_wall.hpp
@@ -25,33 +25,41 @@ using namespace smacc2::default_transition_tags;
 
 // STATE DECLARATION
 //
-// Collision-abort demo, step 1: quarter turn to face the pillar 1.5 m north
-struct StSpinToPillar : smacc2::SmaccState<StSpinToPillar, SmNav2GazeboTest3>
+// Collision-abort demo, step 2: command a drive straight AT the south wall
+// (3.5 m requested, wall ~2 m away and spanning the whole arena). The behavior
+// server's collision checking must abort the goal before impact - the EXPECTED
+// outcome here is EvCbFailure, proving the genuine server-abort path end to
+// end. Success would mean the collision checking failed to intervene.
+struct StDriveAtWall : smacc2::SmaccState<StDriveAtWall, SmNav2GazeboTest3>
 {
   using SmaccState::SmaccState;
 
   // CUSTOM TRANSITION TAGS
   struct NEXT : SUCCESS {};
+  struct COLLISION_ABORT_OK : SUCCESS {};
 
   // TRANSITION TABLE
   typedef mpl::list<
-    Transition<EvCbSuccess<CbSpin, OrNavigation>, StDriveAtPillar, SUCCESS>,
-    Transition<EvCbFailure<CbSpin, OrNavigation>, StDriveAtPillar, ABORT>,
-    Transition<EvKeyPressN<CbDefaultKeyboardBehavior, OrKeyboard>, StDriveAtPillar, NEXT>
+    // expected: collision checking aborts the goal -> failure event
+    Transition<EvCbFailure<CbDriveOnHeading, OrNavigation>, StBackUpSafe, COLLISION_ABORT_OK>,
+    // unexpected but non-fatal: server let the full distance through
+    Transition<EvCbSuccess<CbDriveOnHeading, OrNavigation>, StBackUpSafe, ABORT>,
+    Transition<EvKeyPressN<CbDefaultKeyboardBehavior, OrKeyboard>, StBackUpSafe, NEXT>
   > reactions;
 
   // STATE FUNCTIONS
   static void staticConfigure()
   {
-    // face the pillar at map (-2, 1): quarter turn counter-clockwise from east
-    configure_orthogonal<OrNavigation, CbSpin>(M_PI / 2);
+    // slow approach so the abort is clearly visible
+    configure_orthogonal<OrNavigation, CbDriveOnHeading>(3.5f, 0.1f);
     configure_orthogonal<OrKeyboard, CbDefaultKeyboardBehavior>();
   }
 
   void onEntry()
   {
     RCLCPP_INFO(
-      getLogger(), "StSpinToPillar: onEntry() - collision-abort demo, turning to face the pillar");
+      getLogger(),
+      "StDriveAtWall: onEntry() - driving at the pillar, expecting a collision abort");
   }
 };
