+++ smacc2_sm_reference_library/sm_nav2_gazebo_test_3/include/sm_nav2_gazebo_test_3/superstates/ss_primitive_loop.hpp
@@ -50,7 +50,7 @@ public:
 
   // TRANSITION TABLE
   typedef mpl::list<
-    Transition<EvLoopEnd<StiPrimitiveLoopStart>, StSpinToPillar, ENDLOOP>
+    Transition<EvLoopEnd<StiPrimitiveLoopStart>, StSpinToWall, ENDLOOP>
   > reactions;
 
   static constexpr int total_iterations() { return 3; }
